LRU Cache library for Python
This implementation uses python dictionary and generic
doubly-linked list. Cache is stored in Entry object in the
doubly-linked list, and a dictionary to quickly lookup the cache
entry. The head of the double-linked list stores the recently-used
cache, while the tail stores the less-recently-used cache. Reorganizing
the cache is done by adding cache entry to head and removing the tail
entry.